// src/models/movie.model.ts
interface Movie {
    name: string;
    origin_name: string;
    genres: string;
    country: string;
    director: string;
    release_year: number;
    status: string;
    actor: string;
    description: string;
    episode_total: string;
    duration: string;
}

export type { Movie as default };